# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "contact": {
      "name": "Ploiu",
      "url": "https://github.com/ploiu"
    },
    "title": "Ploiu File Server",
    "description": "a self-hostable file server written in rust",
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/disk": {
      "get": {
        "tags": ["api"],
        "description": "Returns current disk usage information",
        "responses": {
          "200": {
            "description": "info retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/diskInfo"
                }
              }
            }
          },
          "400": {
            "description": "file server is hosted on windows. Accurately determining the disk requires looking at the device id, which is complicated on windows and uses unsafe code. This endpoint will always return a 400 (if auth check passes) and the server is running on windows",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/unauthorizedResponse"
          },
          "500": {
            "description": "failed to retrieve disk info due to an unknown error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          }
        }
      }
    },
    "/api/password": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createPassword"
              }
            }
          }
        },
        "tags": [
          "api"
        ],
        "description": "sets the password to access the api. All non /api endpoints will not work unless a password is set",
        "responses": {
          "201": {
            "description": "password created"
          },
          "400": {
            "description": "password has already been set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "500": {
            "description": "error occurred setting password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          }
        }
      },
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/updatePassword"
              }
            }
          }
        },
        "tags": [
          "api"
        ],
        "description": "updates the password to access the api. All non /api endpoints will not work unless a password is set",
        "responses": {
          "204": {
            "description": "password updated"
          },
          "401": {
            "description": "Failed to set password. Check server logs for details"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "description": "localhost",
      "url": "https://localhost:8000"
    }
  ],
  "components": {
    "schemas": {
...
      "diskInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "the mounted folder name"
          },
          "totalSpace": {
            "type": "integer",
            "format": "int64",
            "description": "the total size in bytes the disk can hold"
          },
          "freeSpace": {
            "type": "integer",
            "format": "int64",
            "description": "the written bytes to the disk"
          }
        }
      },
...
      "basicMessage": {
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "createPassword": {
        "properties": {
          "username": {
            "type": "string",
            "nullable": false
          },
          "password": {
            "type": "string",
            "nullable": false
          }
        }
      },
      "updatePassword": {
        "properties": {
          "oldAuth": {
            "type": "object",
            "properties": {
              "username": {
                "type": "string",
                "nullable": false
              },
              "password": {
                "type": "string",
                "nullable": false
              }
            }
          },
          "newAuth": {
            "type": "object",
            "properties": {
              "username": {
                "type": "string",
                "nullable": false
              },
              "password": {
                "type": "string",
                "nullable": false
              }
            }
          }
        }
      }
    },
    "responses": {
      "unauthorizedResponse": {
        "description": "missing or invalid password",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/basicMessage"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
  "security": [
    {
      "basicAuth": []
    }
  ]
}